Move distributed TimeMonitor in a separate header
